django-starter/config/settings/base.py
"""
config/settings/base.py

Base settings shared across all environments (dev, prod).
Do NOT put secrets here. Use environment variables via .env
"""

import os
from pathlib import Path
from dotenv import load_dotenv

# Load .env file
load_dotenv()

# ─────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────
# Build paths inside the project: BASE_DIR / 'subdir'
# config/settings/base.py → go up 3 levels to reach project root
BASE_DIR = Path(__file__).resolve().parent.parent.parent


# ─────────────────────────────────────────
# SECURITY (overridden per environment)
# ─────────────────────────────────────────
SECRET_KEY = os.environ.get('SECRET_KEY')

DEBUG = False  # always False in base; dev.py sets it to True

ALLOWED_HOSTS = []


# ─────────────────────────────────────────
# APPLICATIONS
# ─────────────────────────────────────────
DJANGO_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',
]

LOCAL_APPS = [
    'apps.accounts',    # Custom user model + auth
    # 'apps.core',      # Add your project-specific apps here
]

INSTALLED_APPS = DJANGO_APPS + LOCAL_APPS


# ─────────────────────────────────────────
# MIDDLEWARE
# ─────────────────────────────────────────
MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'whitenoise.middleware.WhiteNoiseMiddleware',   # serves static files in prod
    'django.contrib.sessions.middleware.SessionMiddleware',
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
]


# ─────────────────────────────────────────
# URLS & WSGI
# ─────────────────────────────────────────
ROOT_URLCONF = 'config.urls'

WSGI_APPLICATION = 'config.wsgi.application'


# ─────────────────────────────────────────
# TEMPLATES
# ─────────────────────────────────────────
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        # Look in the project-level templates/ folder first
        'DIRS': [BASE_DIR / 'templates'],
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
            ],
        },
    },
]


# ─────────────────────────────────────────
# DATABASE
# ─────────────────────────────────────────
# Default to PostgreSQL. Override DATABASE_URL in .env
DATABASES = {
    'default': {
        'ENGINE': 'django.db.backends.postgresql',
        'NAME': os.environ.get('DB_NAME', 'starter_db'),
        'USER': os.environ.get('DB_USER', 'postgres'),
        'PASSWORD': os.environ.get('DB_PASSWORD', ''),
        'HOST': os.environ.get('DB_HOST', 'localhost'),
        'PORT': os.environ.get('DB_PORT', '5432'),
    }
}


# ─────────────────────────────────────────
# CUSTOM USER MODEL
# ─────────────────────────────────────────
# Points to accounts/models.py → CustomUser
AUTH_USER_MODEL = 'accounts.CustomUser'


# ─────────────────────────────────────────
# AUTH REDIRECTS
# ─────────────────────────────────────────
LOGIN_URL = '/accounts/login/'
LOGIN_REDIRECT_URL = '/dashboard/'
LOGOUT_REDIRECT_URL = '/'


# ─────────────────────────────────────────
# PASSWORD VALIDATION
# ─────────────────────────────────────────
AUTH_PASSWORD_VALIDATORS = [
    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
    {'NAME': 'django.contrib.auth.password_validation.CommonPasswordValidator'},
    {'NAME': 'django.contrib.auth.password_validation.NumericPasswordValidator'},
]


# ─────────────────────────────────────────
# INTERNATIONALIZATION
# ─────────────────────────────────────────
LANGUAGE_CODE = 'en-us'
TIME_ZONE = 'Africa/Lagos'      # Change to your timezone
USE_I18N = True
USE_TZ = True


# ─────────────────────────────────────────
# STATIC & MEDIA FILES
# ─────────────────────────────────────────
STATIC_URL = '/static/'
STATICFILES_DIRS = [BASE_DIR / 'static']       # dev: source static files
STATIC_ROOT = BASE_DIR / 'staticfiles'          # prod: collectstatic target

MEDIA_URL = '/media/'
MEDIA_ROOT = BASE_DIR / 'media'                 # uploaded files

# Whitenoise compression for production
STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'


# ─────────────────────────────────────────
# DEFAULT PRIMARY KEY
# ─────────────────────────────────────────
DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'


# ─────────────────────────────────────────
# MESSAGES FRAMEWORK
# ─────────────────────────────────────────
from django.contrib.messages import constants as messages

MESSAGE_TAGS = {
    messages.DEBUG: 'debug',
    messages.INFO: 'info',
    messages.SUCCESS: 'success',
    messages.WARNING: 'warning',
    messages.ERROR: 'error',
}
