django-starter/requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==21.2.0
